co-service-planner/requirements.txt
crewai
crewai-tools
anthropic
openai
pydantic
python-dotenv
